fix: code review remarks
