+++ app/app/admin/layout.tsx
@@ -2,20 +2,10 @@
 
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import { LogOut, Settings, Users } from 'lucide-react';
+import { Users } from 'lucide-react';
 
 import { authClient } from '@/app/lib/auth';
 import { useIsAdmin } from '@/app/lib/role';
-import { Avatar, AvatarFallback, AvatarImage } from '@/app/components/ui/avatar';
-import { Button } from '@/app/components/ui/button';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/app/components/ui/dropdown-menu';
 import { Separator } from '@/app/components/ui/separator';
 import {
   Sidebar,
@@ -33,6 +23,7 @@ import {
 } from '@/app/components/ui/sidebar';
 import { Skeleton } from '@/app/components/ui/skeleton';
 import { ThemeToggle } from '@/app/components/theme-toggle';
+import { UserMenu } from '@/app/components/user-menu';
 
 const sidebarItems = [
   {
@@ -121,58 +112,12 @@ export default function AdminLayout({ children }: { children: React.ReactNode })
             {isSessionPending ? (
               <Skeleton className="h-8 w-8 rounded-full" />
             ) : session ? (
-              <UserMenu name={session.user.name} email={session.user.email} image={session.user.image} />
+              <UserMenu name={session.user.name} email={session.user.email} image={session.user.image} size="sm" />
             ) : null}
           </div>
         </header>
         <main className="flex-1 p-6">{children}</main>
       </SidebarInset>
     </SidebarProvider>
-  );
-}
-
-function UserMenu({ name, email, image }: { name: string; email: string; image?: string | null }) {
-  const handleSignOut = async () => {
-    await authClient.signOut();
-  };
-
-  const initials = name
-    .split(' ')
-    .map((n) => n[0])
-    .join('')
-    .toUpperCase()
-    .slice(0, 2);
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="ghost" className="relative h-8 w-8 rounded-full">
-          <Avatar size="sm">
-            {image && <AvatarImage src={image} alt={name} />}
-            <AvatarFallback>{initials}</AvatarFallback>
-          </Avatar>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-56">
-        <DropdownMenuLabel className="font-normal">
-          <div className="flex flex-col space-y-1">
-            <p className="text-sm leading-none font-medium">{name}</p>
-            <p className="text-muted-foreground text-xs leading-none">{email}</p>
-          </div>
-        </DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem asChild>
-          <Link href="/app/account/settings">
-            <Settings className="mr-2 h-4 w-4" />
-            Settings
-          </Link>
-        </DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={handleSignOut} variant="destructive">
-          <LogOut className="mr-2 h-4 w-4" />
-          Sign out
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
   );
 }
